Add adaptive evaluation script and refactor imports for compatibility
- Introduced `evaluate_adaptive_policy.py` for evaluating LLaDA and Dream models with adaptive policies.
- Added `run_adaptive_eval.sh` script to automate evaluation for both models.
- Refactored import statements in `eval.py`, `eval_llada.py`, and `generate.py` to handle ImportError gracefully.
- Updated softmax precision in `train_adaptive_policy.py` for improved numerical stability.

+++ train_adaptive_policy.py
@@ -107,7 +107,7 @@ def get_transfer_index(
     logits_with_noise = logits
     x0 = torch.argmax(logits_with_noise, dim=-1)
 
-    p = F.softmax(logits.to(torch.float64), dim=-1)
+    p = F.softmax(logits.to(torch.float32), dim=-1)
     x0_p = torch.gather(p, dim=-1, index=x0.unsqueeze(-1)).squeeze(-1)
 
     x0 = torch.where(mask_index, x0, x)
